# Fix `wp_head` performance regression for classic themes