* A performance analysis for WordPress 6.1 has shown that there is a performance regression in `wp_head` due to additional logic focused on FSE, which particularly for classic themes appears to be unnecessary.
* See relevant Trac ticket comments discussing this so far:
    * https://core.trac.wordpress.org/ticket/56467#comment:408
    * https://core.trac.wordpress.org/ticket/56467#comment:411
    * https://core.trac.wordpress.org/ticket/56467#comment:413
* The original implementation of this PR improves median `wp_head` execution time from **16.45ms to 11.92ms (~28% faster)** for classic themes (concretely tested with Twenty Twenty-One), while apparently not missing any actual functionality needed.
* However, based on https://github.com/WordPress/wordpress-develop/pull/3536#issuecomment-1307699049, a more complex part of this implementation was broken out into a separate Gutenberg PR https://github.com/WordPress/gutenberg/issues/45616, so that the rest of this PR can realistically make it into the WordPress 6.1.1 release.
* The latest implementation of this PR still improves median `wp_head` execution time from **16.45ms to 14.78ms (~10% faster)** for the same test environment. This is quite a bit less of an improvement from the original, but it makes sense as with the latest implementation still core's `theme.json` and `theme-i18n.json` files are being parsed. That said, this still brings a worthwhile performance improvement and sets the stage for the follow up work in https://github.com/WordPress/gutenberg/issues/45616 to address the remaining part of the performance issues.
* See [this spreadsheet](https://docs.google.com/spreadsheets/d/1q6PNZ-7uIPJ9MTC3JiXshpNyG7sm28QJOtmavKexsIU/edit) for the test results.

cc @scruffian @oandregal @draganescu 

Trac ticket: https://core.trac.wordpress.org/ticket/56945

---
**This Pull Request is for code review only. Please keep all other discussion in the Trac ticket. Do not merge this Pull Request. See [GitHub Pull Requests for Code Review](https://make.wordpress.org/core/handbook/contribute/git/github-pull-requests-for-code-review/) in the Core Handbook for more details.**