@oandregal 2022-11-02
:wave: I've seen the changes and this is what I can share:

- `get_core_data`: this data needs to be available both for classic and block themes. It loads the default editor configuration (for example, the default color palettes)
- `get_block_data`: this data is necessary for both classic and block themes. It loads style information from blocks.
- `get_theme_data`: this data is required for both classic and block themes. For classic, it takes it from the theme supports and returns it as if it was a theme.json. There may be some opportunities to micro-optimize here, such as do not load the i18n schema if the theme and its parent do not have a theme.json (we'd need to measure how impactful that single change is, that´d be a nice single PR).
- `get_user_data`: this is safe to skip for classic themes. There are edge cases, such as a theme that has theme.json and then it removes it, though I'm inclined to think user styles are safe to ignore in that case. Coincidentally, I've done the same thing in a [prototype](https://github.com/WordPress/gutenberg/pull/45372) I'm working on. It can be extracted as a small, focused PR.

Coincidentally, for the area of "theme.json APIs", I'm compiling a list of tasks I find most pressing at https://github.com/WordPress/gutenberg/issues/45171 I'm very optimistic about how those will impact the performance and code quality. For example, I've just published a [prototype](https://github.com/WordPress/gutenberg/pull/45372) that reduces the total time of a request from 514.44ms to 343.71ms, a 33% performance increase.

Another thing that it'd be useful is to work in these PRs in Gutenberg directly. That way, they can be tested widely and bugs will be raised earlier than the WordPress RC cycles, which is always a lot less stressful and give time to developers to figure out long-term solutions rather than patches for the RC.