@felixarntz 2022-11-02
@oandregal Thank you for providing additional context on why (parts of) these methods are needed for classic themes as well. Based on what you're saying, I still think there are a number of performance optimizations to make here to avoid running FSE-only logic for classic themes. Here are some of the high-level ideas I'm thinking:
* In `get_core_data()`, parsing WP's own `theme.json` is a rather wasteful implementation in itself, more so with the additional translation parsing. Since this file is shipped by core (and thus never changes between core updates), we should find a more efficient solution. For example, how about we generate a fully PHP-ready parsed version including i18n strings of WP's `theme.json` at build time? I see how `theme.json` is certainly more portable than the same data in PHP array format, but we could create an autogenerated version of that file, like a `theme-json.php` file that can then simply be `require`d by PHP, no JSON parsing and i18n schema parsing necessary at runtime.
* In `get_block_data()`, everything looks fine as is to me.
* In `get_theme_data()`, we should simply not parse the i18n schema and try to parse the theme's `theme.json` if it doesn't exist, like you're saying and like this PR already does.
* `get_user_data()` can bail early if in a classic theme, also like you're saying and like this PR already does.

I'll make some updates to the proposed code here accordingly, the main thing though that requires some more work would be to generate some version of a `theme-json.php` file as part of the build process. If we build something like that, we could even share it with e.g. theme developers so they can run it on their `theme.json` files too.

> Felix, would you share the steps you did to test this, so it can be replicated in a variety of environments?

It was a fairly basic process, here's what I did:
* Modify WP core's `wp_head()` function directly, to measure `microtime( true )` before the `do_action( 'wp_head' )` call and after it, then get the difference to know how long the action took.
* I did this with `trunk` and with this PR branch, using a site with Twenty Twenty-One on just the regular basic "Hello world" singular post URL.
* I ran both tests 9 times each (to account for variance) and calculated the median.

If we wanted to dive in more deeply, it would also be interesting to see what the difference would be on a more complex piece of content, e.g. the Gutenberg demo page.