@felixarntz 2022-11-02
@oandregal I've pushed some updates here as mentioned above, please have a look. I have included what the `theme-json.php` file could look like in principle, although of course this would need to be auto-generated by a build process eventually.

One additional concern/question I have: Why is the condition around whether to reuse the previously "cached" theme data only present around parsing the (child) theme's `theme.json` (see https://github.com/WordPress/wordpress-develop/pull/3536/files#diff-d6b86476eed058e7cf8b6e57fa52c4fd75b1f0907e1a9ccb0149528a24f7578bR254)? Are there any problems if we change it to apply to the entire method, similar to how it is for the other `get_*_data()` methods? Otherwise some of that logic is still run on every single method call over and over again. cc @hellofromtonya