@spacedmonkey 2022-11-03
This PR is heavily related - https://github.com/WordPress/wordpress-develop/pull/3556.