@felixarntz 2022-11-08
@spacedmonkey 
> This is hard to code review the code to generate theme-php.php, fix the unit test and coding standards.

No point to review the `theme-json.php` file, as eventually it will be auto-generated. There's also no point in fixing the WPCS etc violations in that file - see the TODO comment and the issue comments above outlining that. :)

I'll work on further polishing here today.

@draganescu @hellofromtonya Any chance I could get your thoughts on the changes and proposed approach from this draft PR? Please make sure to review my above issue comments for context.