@spacedmonkey 2022-11-08
>No point to review the theme-json.php file, as eventually it will be auto-generated. There's also no point in fixing the WPCS etc violations in that file - see the TODO comment and the issue comments above outlining that. :)

I don't think we should merge this without the automatic generation code, in JS grunt. 

I think we should consider making this out to another PR,  where we have more time to review / test.