@felixarntz 2022-11-08
@spacedmonkey 
> I don't think we should merge this without the automatic generation code, in JS grunt.

That's exactly what I'm saying.

> I think we should consider making this out to another PR, where we have more time to review / test.

Sounds good, I've opened https://github.com/WordPress/gutenberg/issues/45616 for that.

Will remove the `theme-json.php` part from here and get this PR ready for 6.1.1.